APi and Client > Error Handling
using FluentValidation and middleware for HTTP request
using axios middle ware to catch errors
added not found and server error page

+++ Application/Activities/Commands/DeleteActivity.cs
@@ -1,4 +1,5 @@
-﻿using Domain;
+﻿using Application.Core;
+using Domain;
 using MediatR;
 using Persistence;
 using System;
@@ -11,21 +12,26 @@ namespace Application.Activities.Commands
 {
     public class DeleteActivity
     {
-        public class Command : IRequest
+        public class Command : IRequest<Result<Unit>>
         {
-            public string Id { get; set; }
+            public required string Id { get; set; }
         }
 
-        public class Handler(AppDBContext context) : IRequestHandler<Command>
+        public class Handler(AppDBContext context) : IRequestHandler<Command, Result<Unit>>
         {
-            public async Task Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
-                var activity = await context.Activities.FindAsync([request.Id], cancellationToken)
-                    ?? throw new Exception("Cannot find the Activity");
-                    
+                var activity = await context.Activities.FindAsync([request.Id], cancellationToken);
+
+                if (activity == null) return Result<Unit>.Failure("Activity Not Found", 404);
+
                 context.Activities.Remove(activity);
 
-                await context.SaveChangesAsync(cancellationToken);
+                var result = await context.SaveChangesAsync(cancellationToken) > 0;
+
+                if(!result) return Result<Unit>.Failure("Failed to delete the activity", 400);
+
+                return Result<Unit>.Success(Unit.Value);
             }
         }
